+++ packages/update-monitor/abi/zkSpace/ZkSpaceVerifier.json
