+++ packages/update-monitor/abi/zkSpace/ZkSpaceZkSync.json
